Created base template for navbar and footer
Created /templates/base.html, a base template with
navigation bar and footer.

Code for navigation bar remains the same,
except that the div tag has been
changed to a nav tag and it has been included in
the header tag of the body. The if control statement
in the navbar should be uncommented once login/logout
has been implemented.

All templates extending the base template should modify
the following blocks (not mandatory):
- {% block title %} for page title (displayed in browser)
- {% block head %} for additional stylesheets or scripts
--- use {{ super() }} in block to call default
- {% block content %} for the main content of the page